DDBTEAM-0101: Moved index event into message handler
