+++ config/packages/messenger.yaml
@@ -51,6 +51,7 @@ framework:
             'App\Message\SearchMessage': async_priority_high
             'App\Message\VendorImageMessage': async_priority_high
             'App\Message\CoverStoreMessage': async_priority_high
+            'App\Message\IndexMessage': async_priority_high
             'App\Message\CoverUserUploadMessage': async_priority_high
             'App\Message\DeleteMessage': async_priority_low
             'App\Message\SearchNoHitsMessage': async_no_hit
