+++ src/MessageHandler/IndexMessageHandler.php
@@ -2,16 +2,15 @@
 
 /**
  * @file
- * Index event subscriber that updates the "search" database table thereby ensuring that the material gets indexed into
- * the search engine.
+ * Index message handler
  */
 
-namespace App\EventSubscriber;
+namespace App\MessageHandler;
 
 use App\Entity\Image;
 use App\Entity\Search;
 use App\Entity\Source;
-use App\Event\IndexReadyEvent;
+use App\Message\IndexMessage;
 use App\Utils\OpenPlatform\Material;
 use Doctrine\DBAL\ConnectionException;
 use Doctrine\DBAL\Exception\UniqueConstraintViolationException;
@@ -19,12 +18,12 @@ use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ManagerRegistry;
 use ItkDev\MetricsBundle\Service\MetricsService;
 use Psr\Log\LoggerInterface;
-use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
 
 /**
- * Class IndexEventSubscriber.
+ * Class IndexMessageHandler.
  */
-class IndexEventSubscriber implements EventSubscriberInterface
+class IndexMessageHandler implements MessageHandlerInterface
 {
     private EntityManagerInterface $em;
     private LoggerInterface $logger;
@@ -32,10 +31,12 @@ class IndexEventSubscriber implements EventSubscriberInterface
     private MetricsService $metricsService;
 
     /**
-     * IndexEventSubscriber constructor.
+     * SearchProcessor constructor.
      *
      * @param EntityManagerInterface $entityManager
      * @param LoggerInterface $informationLogger
+     * @param ManagerRegistry $registry
+     * @param MetricsService $metricsService
      */
     public function __construct(EntityManagerInterface $entityManager, LoggerInterface $informationLogger, ManagerRegistry $registry, MetricsService $metricsService)
     {
@@ -45,28 +46,11 @@ class IndexEventSubscriber implements EventSubscriberInterface
         $this->metricsService = $metricsService;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public static function getSubscribedEvents(): array
-    {
-        return [
-            IndexReadyEvent::NAME => 'onIndexEvent',
-        ];
-    }
-
-    /**
-     * Updated event handler.
-     *
-     * @param IndexReadyEvent $event
-     *
-     * @return void
-     */
-    public function onIndexEvent(IndexReadyEvent $event): void
+    public function __invoke(IndexMessage $message)
     {
         try {
-            $material = $event->getMaterial();
-            $image = $this->getImage($event->getImageId());
+            $material = $message->getMaterial();
+            $image = $this->getImage($message->getImageId());
             $source = (null !== $image) ? $image->getSource() : null;
 
             // If image or source are null something is broken in the data,
@@ -139,7 +123,7 @@ class IndexEventSubscriber implements EventSubscriberInterface
             $this->logger->error('Database Connection Exception', [
                 'service' => 'IndexEventSubscriber',
                 'message' => $exception->getMessage(),
-                'identifiers' => $event->getMaterial()->getIdentifiers(),
+                'identifiers' => $message->getMaterial()->getIdentifiers(),
             ]);
         } catch (\Exception $exception) {
             $this->logger->error('Index Exception', [
