+++ src/MessageHandler/SearchMessageHandler.php
@@ -8,18 +8,18 @@
 namespace App\MessageHandler;
 
 use App\Entity\Source;
-use App\Event\IndexReadyEvent;
 use App\Exception\MaterialTypeException;
 use App\Exception\OpenPlatformAuthException;
 use App\Exception\OpenPlatformSearchException;
+use App\Message\IndexMessage;
 use App\Message\SearchMessage;
 use App\Service\OpenPlatform\SearchService;
 use App\Utils\Types\VendorState;
 use Doctrine\ORM\EntityManagerInterface;
 use Psr\Log\LoggerInterface;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Messenger\Exception\UnrecoverableMessageHandlingException;
 use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
+use Symfony\Component\Messenger\MessageBusInterface;
 
 /**
  * Class SearchMessageHandler.
@@ -27,7 +27,7 @@ use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
 class SearchMessageHandler implements MessageHandlerInterface
 {
     private EntityManagerInterface $em;
-    private EventDispatcherInterface $dispatcher;
+    private MessageBusInterface $bus;
     private LoggerInterface $logger;
     private SearchService $searchService;
 
@@ -35,14 +35,14 @@ class SearchMessageHandler implements MessageHandlerInterface
      * SearchProcessor constructor.
      *
      * @param EntityManagerInterface $entityManager
-     * @param EventDispatcherInterface $eventDispatcher
+     * @param MessageBusInterface $bus
      * @param LoggerInterface $informationLogger
      * @param SearchService $searchService
      */
-    public function __construct(EntityManagerInterface $entityManager, EventDispatcherInterface $eventDispatcher, LoggerInterface $informationLogger, SearchService $searchService)
+    public function __construct(EntityManagerInterface $entityManager, MessageBusInterface $bus, LoggerInterface $informationLogger, SearchService $searchService)
     {
         $this->em = $entityManager;
-        $this->dispatcher = $eventDispatcher;
+        $this->bus = $bus;
         $this->logger = $informationLogger;
         $this->searchService = $searchService;
     }
@@ -116,14 +116,14 @@ class SearchMessageHandler implements MessageHandlerInterface
                 'imageId' => $message->getImageId(),
             ]);
         } else {
-            $event = new IndexReadyEvent();
-            $event->setIs($message->getIdentifier())
+            $indexMessage = new IndexMessage();
+            $indexMessage->setIdentifier($message->getIdentifier())
                 ->setOperation($message->getOperation())
                 ->setVendorId($message->getVendorId())
                 ->setImageId($message->getImageId())
                 ->setMaterial($material);
 
-            $this->dispatcher->dispatch($event, $event::NAME);
+            $this->bus->dispatch($message);
         }
 
         // Free memory.
